@@ -14,13 +14,13 @@ import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.IBinder;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.android.material.snackbar.Snackbar;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -210,7 +210,7 @@ public class MyMediaPlayer extends AppCompatActivity {
     protected void onCreate(final Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_media_player);
-        final Toolbar toolbar = findViewById(R.id.toolbar_notebooks);
+        final Toolbar toolbar = findViewById(R.id.toolbar_media_player);
         setSupportActionBar(toolbar);
 
 //        Intent serviceIntent = new Intent(getApplicationContext(), PlayerService.class);
@@ -311,6 +311,7 @@ public class MyMediaPlayer extends AppCompatActivity {
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
 
+        super.onActivityResult(requestCode, resultCode, data);
         if (checkPermissionWRITE_EXTERNAL_STORAGE(this)) {
 
             if (requestCode == FILE_SELECT_CODE && resultCode == RESULT_OK && data != null) {
